+++ doc/release_notes.rst
@@ -31,6 +31,8 @@ API additions in X.X.0:
 
 Introduced ``install_weak_deps`` :doc:`configuration <conf_ref>` option.
 
+Implemented ``strict`` :doc:`configuration <conf_ref>` option.
+
 API deprecations in X.X.0:
 
 * ``dnf.callback.LoggingTransactionDisplay`` is deprecated now. It was considered part of API despite the fact that it has never been documented. Use :class:`dnf.callback.TransactionProgress` instead.